Adds installer options for meta-networking. Plus, it's pretty. [#39 state:resolved]
When installing, the installer now uses the result of trying to download
    - http://www.example.com/peopleaggregator.txt
    - http://example.com/peopleaggregator.txt
    - http://some-random-domain.example.com/peopleaggregator.txt
to determine if the server is compatible with meta-networking. If all three
downloads succeed, the installer presents the user with an option on the
following page to enable meta-networking. If any of the downloads fail,
it is assumed that meta-networking is not supported on this server due to the
install being on a subdomain.

Also, the installer tries to guess the domain_prefix value. If it guesses
incorrectly, the user can edit its value.