Adds confirmation to 'Remove Picture' prompt. [#58 state:resolved]
This is necessary so the form can submit the right way and not have
POST interfereing with GET. It used to be quite hairy, but I think this
fix is elegant and gets the problem solved in a better way.

Also, the user can now upload a new photo and remove the old photo, and
the expected behavior occurs: the user is left with the new profile pic.