Users installing PA can now opt to use a databes from an existing PA install. They will only be allowed to enter a username and password which can connect to the existing database. [#57 state:resolved]
Upon hitting next, the install process will ensure the user can log in
and has access to the database, and that the database is not empty.
If these conditions are met, then the installer assumes the db is filled
with an existing PA install, and then runs the update script packaged with
PA. In this way, brand-new PA installs can be set up to use the database from
older ones.