Fixes some things in the Image uploading process.
Since we upgraded jQuery in commit dafb028, it broke the image uploading.
I now suspect this is due to the difference in handling CSS3 selectors
between jQuery versions. In older versions, $(":input[@name=tom]") selects
all <input> tags with name="tom". In the latest version, however, the
attribute selector changed to $(":input[name=tom]") to select the same
thing. Consequently, the image uploader was not uploading the image
because the pre-submit method to remove _just the submit button_ now removed
_*all* input fields_. Removing the at-sign seems to have rectified the problem.

So, while here, I figured an upgrade to the jQuery form plugin was in order.

Testing confirms that image uploading works in Firefox and Chrome.